Fix rubocop workflow properly for push trigger (#7)
Instead of not running on the push trigger for Dependabot branches,
which is not fully effective because it still fails on main when a
merge is done by Dependabot due to `@dependabot merge` or similar
command, this just specifies the required permissions in the
workflow file, as is done e.g. in automatically generated CodeQL
workflow files and as suggested in #6.

I tested this in https://github.com/EliahKagan/hello-world. See
https://github.com/EliahKagan/hello-world/pull/14 including
https://github.com/EliahKagan/hello-world/commit/d565b8475c7bda154aec5c2e1b46a3a12daed70d,
https://github.com/EliahKagan/hello-world/pull/15, and
https://github.com/EliahKagan/hello-world/pull/16.